account

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/channel/LabelServiceImpl.java
@@ -6,10 +6,7 @@ import com.zjtelcom.cpct.bean.RespInfo;
 import com.zjtelcom.cpct.common.Page;
 import com.zjtelcom.cpct.dao.channel.*;
 import com.zjtelcom.cpct.domain.channel.*;
-import com.zjtelcom.cpct.dto.channel.LabelAddVO;
-import com.zjtelcom.cpct.dto.channel.LabelGrpVO;
-import com.zjtelcom.cpct.dto.channel.LabelVO;
-import com.zjtelcom.cpct.dto.channel.LabelValueVO;
+import com.zjtelcom.cpct.dto.channel.*;
 import com.zjtelcom.cpct.service.BaseService;
 import com.zjtelcom.cpct.service.channel.LabelService;
 import com.zjtelcom.cpct.util.BeanUtil;
@@ -125,9 +122,9 @@ public class LabelServiceImpl extends BaseService implements LabelService {
     }
 
     @Override
-    public Map<String,Object> editLabel(Long userId, Label editVO) {
+    public Map<String,Object> editLabel(Long userId, LabelEditVO editVO) {
         Map<String,Object> result = new HashMap<>();
-        Label label = labelMapper.selectByPrimaryKey(editVO.getInjectionLabelId());
+        Label label = labelMapper.selectByPrimaryKey(editVO.getLabelId());
         if (label==null){
             result.put("resultCode",CODE_FAIL);
             result.put("resultMsg","标签信息不存在");
@@ -151,6 +148,11 @@ public class LabelServiceImpl extends BaseService implements LabelService {
             result.put("resultMsg","标签信息不存在");
             return result;
         }
+        if (label.getScope().equals(1)){
+            result.put("resultCode",CODE_FAIL);
+            result.put("resultMsg","大数据标签不能删除");
+            return result;
+        }
         labelMapper.deleteByPrimaryKey(labelId);
         result.put("resultCode",CODE_SUCCESS);
         result.put("resultMsg","添加成功");
@@ -158,20 +160,12 @@ public class LabelServiceImpl extends BaseService implements LabelService {
     }
 
     @Override
-    public Map<String,Object> getLabelList(Long userId, Map<String, Object> params, Integer page, Integer pageSize) {
+    public Map<String,Object> getLabelList(Long userId, String labelName, String labelCode, Integer scope, String conditionType, String fitDomain, Integer page, Integer pageSize) {
         Map<String,Object> result = new HashMap<>();
         List<LabelVO> voList = new ArrayList<>();
         List<Label> labelList = new ArrayList<>();
-            String labelName = null;
-            String fitDomain = null;
-            if (params.get("labelName")!=null){
-                labelName = params.get("labelName").toString();
-            }
-            if (params.get("fitDomain")!=null){
-                fitDomain = params.get("fitDomain").toString();
-            }
             PageHelper.startPage(page,pageSize);
-            labelList = labelMapper.findByParam(labelName,fitDomain);
+            labelList = labelMapper.findLabelList(labelName,fitDomain,labelCode,scope,conditionType);
             Page pageInfo = new Page(new PageInfo(labelList));
             for (Label label : labelList){
                 LabelVO vo = ChannelUtil.map2LabelVO(label);
